# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticketing System API",
        "description": "API documentation for Ticketing System",
        "contact": {
            "email": "support@ticketingsystem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/rate-management/rates": {
            "post": {
                "tags": [
                    "Rate Management"
                ],
                "summary": "Create a new rate",
                "description": "Store a newly created rate in storage.",
                "operationId": "ec213c09a0003a9956d0117dc58ba099",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Rate name",
                                        "type": "string",
                                        "example": "Regular Ticket"
                                    },
                                    "description": {
                                        "description": "Rate description",
                                        "type": "string",
                                        "example": "Standard admission ticket"
                                    },
                                    "price": {
                                        "description": "Rate price",
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "status": {
                                        "description": "Rate status",
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rate created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/options/rates": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Get active rates for dropdown",
                "description": "Get rates for dropdown selection.",
                "operationId": "bed32fc6795508299a63eb1329ffeadf",
                "responses": {
                    "200": {
                        "description": "List of active rates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Regular Ticket"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 500
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}